moving RootTreeReader to module Framework/Utils
