+++ Detectors/TPC/workflow/src/DigitReaderSpec.cxx
@@ -15,7 +15,7 @@
 
 #include "DigitReaderSpec.h"
 #include "Headers/DataHeader.h"
-#include "Framework/RootTreeReader.h"
+#include "Utils/RootTreeReader.h"
 #include <memory> // for make_shared, make_unique, unique_ptr
 
 using namespace o2::framework;
