+++ Framework/Utils/CMakeLists.txt
@@ -24,6 +24,7 @@ set(SRCS
 
 set(HEADERS
         include/Utils/Utils.h
+        include/Utils/RootTreeReader.h
         )
 
 ## TODO: feature of macro, it deletes the variables we pass to it, set them again
@@ -43,4 +44,15 @@ O2_FRAMEWORK_WORKFLOW(
         WORKFLOW_NAME "test_DPLOutputTest"
         DETECTOR_BUCKETS DPLUtils_bucket
         SOURCES src/Utils.cxx test/test_DPLOutputTest.cxx test/DPLOutputTest.cxx
+)
+
+set(TEST_SRCS
+      test/test_RootTreeReader.cxx
+   )
+
+O2_GENERATE_TESTS(
+  MODULE_LIBRARY_NAME ${LIBRARY_NAME}
+  BUCKET_NAME ${BUCKET_NAME}
+  TEST_SRCS ${TEST_SRCS}
+  TIMEOUT 30
 )
